Test: ADD SplittableRandom
