+++ java/java-api/java-api/src/main/java/me/staek/stream/scenario/Stream_Ex1.java
