+++ java/java-api/java-api/src/main/java/me/staek/stream/scenario/Stream_Ex2.java
