+++ java/java-api/java-api/src/main/java/me/staek/stream/scenario/Stream_Scenario.java
@@ -1,10 +1,11 @@
 package me.staek.stream;
 
+import me.staek.stream.vo.Study;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 public class Stream_Scenario {
     public static void main(String[] args) {
