+++ java/java-api/java-api/src/main/java/me/staek/stream/vo/Study.java
